src/requirements.txt
absl-py==0.13.0
argon2-cffi==20.1.0
asgiref==3.5.2
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
clr-loader==0.1.7
colorama==0.4.4
cryptography==37.0.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-db-views==0.1.4
django-extensions==3.2.1
django-filter==22.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
entrypoints==0.3
filelock==3.8.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.33.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h11==0.13.0
h5py==3.1.0
idna==3.2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mysql-connector-python==8.0.30
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.0
pandas==1.4.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
pipenv==2022.9.8
platformdirs==2.5.2
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pythonnet @ git+https://github.com/pythonnet/pythonnet@987b2eec097b9c83920ea8285d5a97d6a19064b5
pytz==2022.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-surprise==1.1.1
scipy==1.7.1
selenium==4.3.0
Send2Trash==1.7.1
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==5.0.5
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==3.7.4.3
tzdata==2022.5
urllib3==1.26.6
virtualenv==20.16.5
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wsproto==1.1.0
